mysite/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-debug-toolbar==3.2.4
django-filter==21.1
django-grappelli==3.0.2
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
p5py==1.0.0
pep517==0.12.0
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
zipp==3.7.0
